Simplified the FeedTable save method
`getArrayCopy` did just what was being done by manually creating the
array from the FeedModel object. So it didn't make sense, on the surface
of it, to keep going this way. So I refactored to use `getarrayCopy``
instead. But I've left the original code in there as a suitable example.